backend/requirements.txt
# backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23

# Environment
python-dotenv==1.0.0

# AI/ML Libraries
openai==1.3.0
openai-whisper==20231117
torch==2.1.0
torchaudio==2.1.0

# Audio/Video Processing
ffmpeg-python==0.2.0

# Document Generation
python-docx==0.8.11

# Utilities
requests==2.31.0
Werkzeug==2.3.7

# Production
gunicorn==21.2.0

# Development
pytest==7.4.3
pytest-flask==1.3.0
psutil==5.9.6
